Groups and collections 1/3
This is the first out of a three parts work.
In this part we use Collections instead of GroupObject to handle group elements.

I'm not asking for review (yet), but if anyone wants to comment on anything, the
sooner the better, so feel fre.

Coming next:
(1) Allow collection of type "group"
(2) Allow creation of groups from collection

I will update the patch description once I get those other parts finished.
But the final commit will likely be done in separated patches.

Reviewers: sergey, mont29

Subscribers: brecht

Differential Revision: https://developer.blender.org/D2892