DRW: Add own opengl context to DrawManager.
We create a new offscreen render context (without a window attached as allowed by ogl > 3.0) and use it to draw everything in the draw manager.
This require the context to be setup right. The output Framebuffer need to be created under the drawmanager context (because FBOs are not shared).
Same goes for deletion. Although, fix for proper deletion will come in another commit.

This should fix the MultiWindow/Threaded Render issue. Now every single batch drawn by the draw manager is declared and drawn by the same context.
So the VAO problem is gone and only one VAO per batch is needed.

VAO deletion is also a concern that still needs to be addressed correctly.

We still have issues with ownership and freeing