DRW: Fix DRW_draw_select_loop and DRW_draw_depth_loop
Depth of selection offscreen buffer was uninitialized and selection was behaving badly.

DRW_draw_depth_loop needed to be updated to use the drawmanager context but since the reading is done is quite some places outside of DRW, I prefered to still copy the depth to the backbuffer until we update all operators to use offscreen buffers.