Move render depsgraph to where it belongs: RenderLayer
This fixes compositing for Eevee as well as lay the ground work to fix Cycles
objects iteration in blender 2.8.

Note: Work in progress, the following is not working:
* Cycles
** see "#ifdef TODO" in blender_session.cpp

* Sequencer
** pipeline.c calls BKE_sequencer_new_render_data() early on, and assumes an eval_ctx from that moment on

* Point cache texture (RE_sample_material_init)
** needs eval_ctx :/

What is not handled here: Get rid of view layer in EvaluationContext.
We should store ViewLayer in RenderLayer. The patch is massive as it is though.