From review: Move depsgraph away from render layer and change render loop
Create depsgraph as part of the render loop and pass a render layer for the
render_to_image. Optionally we could create a single layered render_result
but why would we?

Depsgraph is created from scratch prior to rendering each frame.
So I got rid of most of the partial update calls we had during the render
pipeline.

Also, it's past the time to remove Blender Internal.

Note 1: Cycles doesn't build.

Note 2: Cycles seems to rely on an "empty" render function, where all the
required settings are passed to the update function instead. I would like
to have this cleared out because draw manager is assuming everything is
passed to render_to_image, but we need to have an API that makes sense
to all cases.