View Navigation: 2D viewport panning
This option helps to keep the working area under control when working
with brushes, especially if this feature is used with rotate around
selection enabled. Now, panning speed does not depend on zoom level and
it behaves more similarly to a 2D environment.

The option certainly needs a better name and description, and it could
be a good idea to hide and disable it if rotate around selection is
disabled.

Mode info: D3931