added a new live remesh option.
If enabled, the modifier will be continously invoked, as usual. If disabled,
it will return the last automatically cached mesh, so basically the remesh
is "frozen", but still nondestructive / not applied yet.
Note, doesnt work yet in sculpt mode. There you may still need to apply
the modifier to update the base mesh. Or you can disable the modifier in
the stack to work on the base mesh, too. A third alternative is having
live remesh enabled with the modifier enabled, but this will cost some
sculpting performance and works at the moment only in material and rendered
mode.