GPencil: Improves Dot Texture rotation
Before when you use textures in a Dot stroke, the texture did not rotate with the stroke and always kept locked to North (0, 1).

Now, the rotation of the texture is calculated using the previous point to determine the direction of the stroke and rotate the texture following the drawing path.