Mesh Filter: Multithreaded initialization and masked node filtering
This should improve performance a lot, making the tool usable on several
millions of vertices. If the mesh is partially masked, only the unmasked
nodes are going to be processed, so it is possible to isolate parts of a
high poly mesh using the mask and work with them in real time.

I also changed the random displacement initialization to avoid
allocating a huge amount of memory (it wasn't really necessary)