Removed Cloth Cache System
This has been done to minimize the possible areas of bugs when adding the remeshing steps, current point cache system may not be compatible.

Python file for the UI has cache related lines commented.
C files has a macro in BKE_cloth.h which can be set to 1 to enable caching again.