Cloth: Initial step towards remeshing
Currently, the mesh is triangulated. This is a necessary step in the remeshing pipeline and it also serves as a means of testing the functionality of the recent change wherein the cloth modifier has been changed to an applyModifier() based modifier.

Mesh generated by cloth_remeshing_step() needs to be sent to clothModifier_do for the changes to reflect in the viewport.