Cloth: Store previous mesh for remeshing step
Final puzzle to connect the remeshing step with the modifier
Currently a hack, but seems like the best possible method for now, will need to interfere with code that calls the modifier stack otherwise. This method uses more memory.
Now can finally proceed with the actual implementation of the paper