Outliner: Fix walk select error on open leaf elements
In the outliner, leaf elements like light or material data do not
show a disclosure triangle. However, they can still be toggled
with a click or the openclose operator. This led to the walk select
attempting to walk down another level in the tree.

This could be fixed in the toggle open/close functions as well, but
it seems to address it in walk select seemed simpler as this is
the only area affected.