Outliner: Draw constraint icons
Draws constraint icons in the outliner for bones and objects. The
only icon that has an issue is the Action constraint icon, because
it doesn't have a specific constraint icon, so it is drawn white
instead of blue.

At first I tried to replicate the code for modifier icons, then
I found a solution using `te->directdata` that worked for both
objects and bones. It works well enough, not sure if that simpler
solution should be used for modifiers too.