Outliner: Refine synced selection
Any selection event that uses ED_object_base_select will now
mark the outliners as being dirty. When an outliner draws and is
dirty with synced selection enabled, it will sync its selection
state with the view layer.

More work still needs to be done. This may not be the best way
to implement it (selections from scripts won't be tracked). Also,
selection from the outliner isn't perfect.