Cloth: start implementation of updating cloth with new vertices
Idea is to flag all the newly created vertices. After all of the remeshing, create a new Cloth and add all the flagged vertices as well, switch out old Cloth for new Cloth. This is done because it seems harder to update the Cloth every time a new vertex is formed, would be easier to flag those vertices in BMesh and update the Cloth after the fact.
Currently added support to find the new ClothVertex's values.
Some design via comments added.