Eyedropper: Support datablock picking in outliner
This adds support for picking datablocks from the outliner with the
eyedropper tool.

This requires changing the region the UI tooltip is drawn in by
tracking when the mouse moves to a different area on the screen.
By doing this the draw callback can be reset for each new area the
eyedropper enters.