Custom Profile Editing: Added a profile editor widget
This ProfileWidget and related drawing and interface code mirrors much of the
functionality of the CurveMapping code, but with the important different that
it's evaluation results in a 2D result, the X and Y of a position along the
profile generated in the widget. Because of this fundamental change to the
purpose of the widget, it made sense to split it into a new widget rather
than increasing the complexity of both areas.

This is incomplete and buggy at the moment, but it lays the groundwork for
future changes that I'll work on in the coming days, and it does compile.