USD: Initial support for exporting animated data
Mesh and transform animation are now written when passing `animation=True`
to the export operator. There is no inspection of whether an object is
actually animated or not, so this will likely export too much information
(each mesh is written for each frame in the scene).

The administration of which timecode to use for the export is left to the
file-format-specific concrete subclasses of `AbstractHierarchyIterator`;
the abstract iterator itself doesn't know anything about the passage of
time.