ProfileWidget: Added preset selection and reverse button.
The preset selection selects from a few examples currently which are just
defined in the C code manually for now. The preset selector is drawn as part
of the ProfileWidget rather than by the bevel modifier.

The reverse function flips all of the points across the y = 1 - x centerline.
It will act as the manual control over the orientation of the profile. It will
probably need a new icon eventually.

This commit also includes various other changes including variable name changes,
updated comments, and cleanup.