Custom bevel profiles: Merged calculate_profile and calculate_profile_custom.
Now that I'm using the custom profile for all bevel cases, I don't need to have
separate code that uses the custom bevel profiles, so I only need one function.

This also changes the "reversed" parameter to flip the order of the calculated
profile so it doesn't have to be done later with get_profile_point indices.