Outliner: Fix tree building not setting parents properly
When the scenes view tree was built, the parent for view layers
and objects were not set properly (either NULL or incorrect
parent), causing walk select to fail. Now the proper parents are
set.

Not sure if there was a reason why the wrong parent was set for
view layers before.